some more progress on the Callstate->State branch
The problem wasn't that Mem.alloc could fail.  Rather, Coq needed a way to
name the existential terms in the result of Mem.alloc, so that it could
propagate these terms to other goals.  This is fairly straightforward
(using "fst" and "snd"), but Coq couldn't figure this out on its own.

Now stuck proving that the initial state after invoking the function is
correct -- and it seems like this is the place where we need to prove
the correctness of the initialization code (zeroing out A and X, etc).

Maybe this means we need to move "star_refl" further down, and expand
a few more steps of "star" using "star_step".