checkpoint some progress on label-finding
The new find_label_code lemma proves the existence of intermediate
instructions, so we can use it easier with "destruct".

Still need to line up the arguments for destruct in transl_step.

Also need to go back and prove the stupid skipn_middle lemma, which
for now is just assumed correct with "Admitted".