start adding ld_mem
For now ld_mem is still disabled in filter.  There's no need for
an additional variable reg_mem, as the scratch memory is on the
stack and use of a stack pointer is sufficient.

Not sure what's the best plan for initializing the scratch memory:

1) zeroize the scratch memory first, as BSD does; or

2) leave the content as undef, and enhance the filter to reject
   bpf programs that read a location before write; this is what
   Linux does.

Ideally, 1) shouldn't impose a performance penalty---the compiler
should be able to optimize away stores to stack if the values are
never read.  It should also be easier to prove.