bounded stack useage theorem
Actually we don't really need to extend CompCert to get the bounded
stack guarantee.  Just add a filter in the end, which checks the
assembly code from CompCert, particularly, the size in the pseudo
instruction `Pallocframe`, and simply rejects code that has a too
large stack size.

We can now move Theorem 10 in the paper from discussion to design.